Distributed manual tests
